+++ lncfg/invoices.go
@@ -1,7 +1,5 @@
 package lncfg
 
-import "fmt"
-
 const (
 	// DefaultHoldInvoiceExpiryDelta defines the number of blocks before the
 	// expiry height of a hold invoice's htlc that lnd will automatically
@@ -41,20 +39,6 @@ const (
 //nolint:lll
 type Invoices struct {
 	HoldExpiryDelta uint32 `long:"holdexpirydelta" description:"The number of blocks before a hold invoice's htlc expires that the invoice should be canceled to prevent a force close. Force closes will not be prevented if this value is not greater than DefaultIncomingBroadcastDelta."`
-
-	BlindedPaths BlindedPaths `group:"blinding" namespace:"blinding"`
-}
-
-// BlindedPaths holds the configuration options for blinded paths added to
-// invoices.
-//
-//nolint:lll
-type BlindedPaths struct {
-	MinNumRealHops           uint8   `long:"min-num-real-hops" description:"The minimum number of real hops to include in a blinded path. This doesn't include our node, so if the minimum is 1, then the path will contain at minimum our node along with an introduction node hop. If it is zero then the shortest path will use our node as an introduction node."`
-	NumHops                  uint8   `long:"num-hops" description:"The number of hops to include in a blinded path. This doesn't include our node, so if it is 1, then the path will contain our node along with an introduction node or dummy node hop. If paths shorter than NumHops is found, then they will be padded using dummy hops."`
-	MaxNumPaths              uint8   `long:"max-num-paths" description:"The maximum number of blinded paths to select and add to an invoice."`
-	PolicyIncreaseMultiplier float64 `long:"policy-increase-multiplier" description:"The amount by which to increase certain policy values of hops on a blinded path in order to add a probing buffer."`
-	PolicyDecreaseMultiplier float64 `long:"policy-decrease-multiplier" description:"The amount by which to decrease certain policy values of hops on a blinded path in order to add a probing buffer."`
 }
 
 // Validate checks that the various invoice config options are sane.
@@ -70,24 +54,6 @@ func (i *Invoices) Validate() error {
 			"delta: %v, accepted hold invoices will force close "+
 			"channels if they are not canceled manually",
 			i.HoldExpiryDelta, DefaultIncomingBroadcastDelta)
-	}
-
-	if i.BlindedPaths.MinNumRealHops > i.BlindedPaths.NumHops {
-		return fmt.Errorf("the minimum number of real hops in a " +
-			"blinded path must be smaller than or equal to the " +
-			"number of hops expected to be included in each path")
-	}
-
-	if i.BlindedPaths.PolicyIncreaseMultiplier < 1 {
-		return fmt.Errorf("the blinded route policy increase " +
-			"multiplier must be greater than or equal to 1")
-	}
-
-	if i.BlindedPaths.PolicyDecreaseMultiplier > 1 ||
-		i.BlindedPaths.PolicyDecreaseMultiplier < 0 {
-
-		return fmt.Errorf("the blinded route policy decrease " +
-			"multiplier must be in the range (0,1]")
 	}
 
 	return nil
