+++ sample-lnd.conf
@@ -1649,41 +1649,6 @@
 ; enough to prevent force closes.
 ; invoices.holdexpirydelta=12
 
-; The minimum number of real (non-dummy) blinded hops to select for a blinded
-; path. This doesn't include our node, so if the maximum is 1, then the
-; shortest paths will contain our node along with an introduction node hop.
-; invoices.blinding.min-num-real-hops=1
-
-; The number of hops to include in a blinded path. This does not include
-; our node, so if is is 1, then the path will at least contain our node along
-; with an introduction node hop. If it is 0, then it will use this node as
-; the introduction node. This number must be greater than or equal to the
-; the number of real hops (invoices.blinding.min-num-real-hops). Any paths
-; shorter than this number will be padded with dummy hops.
-; invoices.blinding.num-hops=2
-
-; The maximum number of blinded paths to select and add to an invoice.
-; invoices.blinding.max-num-paths=3
-
-; The amount by which to increase certain policy values of hops on a blinded
-; path in order to add a probing buffer. The higher this multiplier, the more
-; buffer is added to the policy values of hops along a blinded path meaning
-; that if they were to increase their policy values before the blinded path
-; expires, the better the chances that the path would still be valid meaning
-; that the path is less prone to probing attacks. However, if the multiplier
-; is too high, the resulting buffered fees might be too much for the payer.
-; invoices.blinding.policy-increase-multiplier=1.1
-
-; The amount by which to decrease certain policy values of hops on a blinded
-; path in order to add a probing buffer. The lower this multiplier, the more
-; buffer is added to the policy values of hops along a blinded path meaning
-; that if they were to increase their policy values before the blinded path
-; expires, the better the chances that the path would still be valid meaning
-; that the path is less prone to probing attacks. However, since this value
-; is being applied to the MaxHTLC value of the route, the lower it is, the
-; lower payment amount will need to be.
-; invoices.blinding.policy-decrease-multiplier=0.9
-
 [routing]
 
 ; DEPRECATED: This is now turned on by default for Neutrino (use
@@ -1697,6 +1662,40 @@
 ; seen as being live from it's PoV.
 ; routing.strictgraphpruning=false
 
+; The minimum number of real (non-dummy) blinded hops to select for a blinded
+; path. This doesn't include our node, so if the maximum is 1, then the
+; shortest paths will contain our node along with an introduction node hop.
+; routing.blinding.min-num-real-hops=1
+
+; The number of hops to include in a blinded path. This does not include
+; our node, so if is is 1, then the path will at least contain our node along
+; with an introduction node hop. If it is 0, then it will use this node as
+; the introduction node. This number must be greater than or equal to the
+; the number of real hops (invoices.blinding.min-num-real-hops). Any paths
+; shorter than this number will be padded with dummy hops.
+; routing.blinding.num-hops=2
+
+; The maximum number of blinded paths to select and add to an invoice.
+; routing.blinding.max-num-paths=3
+
+; The amount by which to increase certain policy values of hops on a blinded
+; path in order to add a probing buffer. The higher this multiplier, the more
+; buffer is added to the policy values of hops along a blinded path meaning
+; that if they were to increase their policy values before the blinded path
+; expires, the better the chances that the path would still be valid meaning
+; that the path is less prone to probing attacks. However, if the multiplier
+; is too high, the resulting buffered fees might be too much for the payer.
+; routing.blinding.policy-increase-multiplier=1.1
+
+; The amount by which to decrease certain policy values of hops on a blinded
+; path in order to add a probing buffer. The lower this multiplier, the more
+; buffer is added to the policy values of hops along a blinded path meaning
+; that if they were to increase their policy values before the blinded path
+; expires, the better the chances that the path would still be valid meaning
+; that the path is less prone to probing attacks. However, since this value
+; is being applied to the MaxHTLC value of the route, the lower it is, the
+; lower payment amount will need to be.
+; routing.blinding.policy-decrease-multiplier=0.9
 
 [sweeper]
